Add CircleDraw in Clojure-Seesaw
It was harder than expected although adding the undo/redo functionality
was easier. Whereas the state was encapsulated in the CircleDrawCanvas
class in the Java7-Swing solution, here it is global. I found no other
satisfying solution but I think it is idiomatic (state should not be
hidden and be centralized).