Improve Cells for Java7-Swing
The code is now closer to Scala's SCells example. The code
is simplified especially the Model class. Bonus: the error
message doesn't overwrite the erroneous input anymore.

The array of the DefaultTableModel is now responsible for
backing/keeping the user input in the respective cells.
When the user edits a cell he will see the content of
the DefaultTableModel's array for that cell (the shadow
table). Otherwise,the value of our cells array is shown
in the cell (see the overridden getCellRenderer method
in Cells).