Delete FreePascal implementation
I decided to delete this contribution since it is incomplete and I did
not find a way to contact the author. It can always be recreated
for the git history.

I did not find a way to contact the author besides his Github profile
where he wasn't active for the last couple of years. Otherwise, I would
have asked him to please transfer the implementation into its own repo.